+++ ballast-repository/src/androidMain/kotlin/com/copperleaf/ballast/repository/AndroidBallastRepository.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.viewModelScope
 import com.copperleaf.ballast.BallastViewModel
 import com.copperleaf.ballast.BallastViewModelConfiguration
 import com.copperleaf.ballast.build
-import com.copperleaf.ballast.core.FifoInputStrategy
 import com.copperleaf.ballast.internal.BallastViewModelImpl
 import com.copperleaf.ballast.repository.bus.EventBus
 import com.copperleaf.ballast.repository.bus.EventBusEventHandler
@@ -25,12 +24,24 @@ private constructor(
 
     public constructor(
         eventBus: EventBus,
+        config: BallastViewModelConfiguration<Inputs, Any, State>,
+    ) : this(
+        impl = BallastViewModelImpl(config),
+        eventBus = eventBus,
+    )
+
+    @Deprecated(
+        message = "Use configBuilder.withRepository().build() instead.",
+        replaceWith = ReplaceWith("AndroidBallastRepository<Inputs, State>(eventBus = eventBus, config = configBuilder.withRepository().build<Inputs, Any, State>())", "com.copperleaf.ballast.build")
+    )
+    public constructor(
+        eventBus: EventBus,
         configBuilder: BallastViewModelConfiguration.Builder,
     ) : this(
         impl = BallastViewModelImpl(
             config = configBuilder
-                .apply { inputStrategy = FifoInputStrategy() }
-                .build<Inputs, Any, State>()
+                .withRepository()
+                .build()
         ),
         eventBus = eventBus,
     )